+++ president/src/main/java/cards/president/CheckerGamePresidentWithRules.java
@@ -530,8 +530,8 @@ public final class CheckerGamePresidentWithRules {
                         str_.add(c.strength(loadedGameCopy_.isReversed()));
                     }
                     if (!str_.isEmpty()) {
-                        boolean same_ = Numbers.eq(str_.getMinimum(),
-                                str_.getMaximum());
+                        boolean same_ = Numbers.eq(str_.getMinimum().byteValue(),
+                                str_.getMaximum().byteValue());
                         // if (!loadedGameCopy_.keepPlayingCurrentGame()) {
                         // same_ = true;
                         // }
